revert progressbarpainter.
add QColor param to progressbarpainter.paint

transferlistdelegate.cpp determines progress bar color
