+++ src/gui/progressbarpainter.cpp
@@ -39,50 +39,7 @@
 #endif
 
 #include "base/global.h"
-#include "base/bittorrent/torrent.h"
 #include "gui/uithememanager.h"
-
-namespace
-{
-    QHash<BitTorrent::TorrentState, QColor> torrentStateColorsFromUITheme()
-    {
-        struct TorrentStateColorDescriptor
-        {
-            const BitTorrent::TorrentState state;
-            const QString id;
-        };
-
-        const TorrentStateColorDescriptor colorDescriptors[] =
-        {
-            {BitTorrent::TorrentState::Downloading, u"TransferList.Downloading"_s},
-            {BitTorrent::TorrentState::StalledDownloading, u"TransferList.StalledDownloading"_s},
-            {BitTorrent::TorrentState::DownloadingMetadata, u"TransferList.DownloadingMetadata"_s},
-            {BitTorrent::TorrentState::ForcedDownloadingMetadata, u"TransferList.ForcedDownloadingMetadata"_s},
-            {BitTorrent::TorrentState::ForcedDownloading, u"TransferList.ForcedDownloading"_s},
-            {BitTorrent::TorrentState::Uploading, u"TransferList.Uploading"_s},
-            {BitTorrent::TorrentState::StalledUploading, u"TransferList.StalledUploading"_s},
-            {BitTorrent::TorrentState::ForcedUploading, u"TransferList.ForcedUploading"_s},
-            {BitTorrent::TorrentState::QueuedDownloading, u"TransferList.QueuedDownloading"_s},
-            {BitTorrent::TorrentState::QueuedUploading, u"TransferList.QueuedUploading"_s},
-            {BitTorrent::TorrentState::CheckingDownloading, u"TransferList.CheckingDownloading"_s},
-            {BitTorrent::TorrentState::CheckingUploading, u"TransferList.CheckingUploading"_s},
-            {BitTorrent::TorrentState::CheckingResumeData, u"TransferList.CheckingResumeData"_s},
-            {BitTorrent::TorrentState::StoppedDownloading, u"TransferList.StoppedDownloading"_s},
-            {BitTorrent::TorrentState::StoppedUploading, u"TransferList.StoppedUploading"_s},
-            {BitTorrent::TorrentState::Moving, u"TransferList.Moving"_s},
-            {BitTorrent::TorrentState::MissingFiles, u"TransferList.MissingFiles"_s},
-            {BitTorrent::TorrentState::Error, u"TransferList.Error"_s}
-        };
-
-        QHash<BitTorrent::TorrentState, QColor> colors;
-        for (const TorrentStateColorDescriptor &colorDescriptor : colorDescriptors)
-        {
-            const QColor themeColor = UIThemeManager::instance()->getColor(colorDescriptor.id);
-            colors.insert(colorDescriptor.state, themeColor);
-        }
-        return colors;
-    }
-}
 
 ProgressBarPainter::ProgressBarPainter(QObject *parent)
     : QObject(parent)
@@ -93,12 +50,11 @@ ProgressBarPainter::ProgressBarPainter(QObject *parent)
     m_dummyProgressBar.setStyle(fusionStyle);
 #endif
 
-    loadUIThemeResources();
     applyUITheme();
     connect(UIThemeManager::instance(), &UIThemeManager::themeChanged, this, &ProgressBarPainter::applyUITheme);
 }
 
-void ProgressBarPainter::paint(QPainter *painter, const QStyleOptionViewItem &option, const QString &text, const int progress, const BitTorrent::TorrentState torrentState) const
+void ProgressBarPainter::paint(QPainter *painter, const QStyleOptionViewItem &option, const QString &text, const int progress, const QColor color) const
 {
     QStyleOptionProgressBar styleOption;
     styleOption.initFrom(&m_dummyProgressBar);
@@ -116,9 +72,9 @@ void ProgressBarPainter::paint(QPainter *painter, const QStyleOptionViewItem &op
     const bool isEnabled = option.state.testFlag(QStyle::State_Enabled);
     styleOption.palette.setCurrentColorGroup(isEnabled ? QPalette::Active : QPalette::Disabled);
 
-    if (torrentState != BitTorrent::TorrentState::Unknown)
+    if (color != nullptr)
     {
-        styleOption.palette.setColor(QPalette::Highlight, m_stateThemeColors.value(torrentState));
+        styleOption.palette.setColor(QPalette::Highlight, color);
     }
     else if (m_chunkColor.isValid())
     {
@@ -136,9 +92,3 @@ void ProgressBarPainter::applyUITheme()
 {
     m_chunkColor = UIThemeManager::instance()->getColor(u"ProgressBar"_s);
 }
-
-void ProgressBarPainter::loadUIThemeResources()
-{
-    m_stateThemeColors = torrentStateColorsFromUITheme();
-}
-
